Handler restructured. Documentation added
