+++ crates/revm/src/handler/mainnet.rs
@@ -1,234 +1,15 @@
 //! Mainnet related handlers.
 
-pub mod frames;
-pub mod host;
-pub mod main;
-pub mod preexecution;
+mod frame;
+mod main;
+mod validation;
 
-use crate::{
-    interpreter::{return_ok, return_revert, Gas, InstructionResult},
-    primitives::{
-        db::Database,
-        Account, EVMError, Env, Spec,
-        SpecId::{CANCUN, LONDON, SHANGHAI},
-        TransactTo, U256,
-    },
-    Context,
+pub use frame::{
+    create_first_frame, frame_return_with_refund_flag, handle_frame_return, handle_frame_sub_call,
+    handle_frame_sub_create, main_frame_return,
 };
-
-pub fn handle_call_return_with_refund_flag<SPEC: Spec>(
-    env: &Env,
-    call_result: InstructionResult,
-    returned_gas: Gas,
-    refund_enabled: bool,
-) -> Gas {
-    // Spend the gas limit. Gas is reimbursed when the tx returns successfully.
-    let mut gas = Gas::new(env.tx.gas_limit);
-    gas.record_cost(env.tx.gas_limit);
-
-    match call_result {
-        return_ok!() => {
-            gas.erase_cost(returned_gas.remaining());
-            gas.record_refund(returned_gas.refunded());
-        }
-        return_revert!() => {
-            gas.erase_cost(returned_gas.remaining());
-        }
-        _ => {}
-    }
-    // Calculate gas refund for transaction.
-    // If config is set to disable gas refund, it will return 0.
-    // If spec is set to london, it will decrease the maximum refund amount to 5th part of
-    // gas spend. (Before london it was 2th part of gas spend)
-    if refund_enabled {
-        // EIP-3529: Reduction in refunds
-        gas.set_final_refund::<SPEC>()
-    };
-
-    gas
-}
-
-/// Handle output of the transaction
-#[inline]
-pub fn handle_call_return<SPEC: Spec>(
-    env: &Env,
-    call_result: InstructionResult,
-    returned_gas: Gas,
-) -> Gas {
-    handle_call_return_with_refund_flag::<SPEC>(env, call_result, returned_gas, true)
-}
-
-pub(crate) fn deduct_caller_inner<SPEC: Spec>(caller_account: &mut Account, env: &Env) {
-    // Subtract gas costs from the caller's account.
-    // We need to saturate the gas cost to prevent underflow in case that `disable_balance_check` is enabled.
-    let mut gas_cost = U256::from(env.tx.gas_limit).saturating_mul(env.effective_gas_price());
-
-    // EIP-4844
-    if SPEC::enabled(CANCUN) {
-        let data_fee = env.calc_data_fee().expect("already checked");
-        gas_cost = gas_cost.saturating_add(data_fee);
-    }
-
-    // set new caller account balance.
-    caller_account.info.balance = caller_account.info.balance.saturating_sub(gas_cost);
-
-    // bump the nonce for calls. Nonce for CREATE will be bumped in `handle_create`.
-    if matches!(env.tx.transact_to, TransactTo::Call(_)) {
-        // Nonce is already checked
-        caller_account.info.nonce = caller_account.info.nonce.saturating_add(1);
-    }
-
-    // touch account so we know it is changed.
-    caller_account.mark_touch();
-}
-
-/// Main load handle
-#[inline]
-pub fn main_load<SPEC: Spec, EXT, DB: Database>(
-    context: &mut Context<EXT, DB>,
-) -> Result<(), EVMError<DB::Error>> {
-    // the L1-cost fee is only computed for Optimism non-deposit transactions.
-    #[cfg(feature = "optimism")]
-    if env.cfg.optimism && env.tx.optimism.source_hash.is_none() {
-        let l1_block_info =
-            optimism::L1BlockInfo::try_fetch(self.context.evm.db).map_err(EVMError::Database)?;
-
-        // storage l1 block info for later use.
-        self.context.evm.l1_block_info = Some(l1_block_info);
-
-        tx_l1_cost
-    }
-
-    // load coinbase
-    // EIP-3651: Warm COINBASE. Starts the `COINBASE` address warm
-    if SPEC::enabled(SHANGHAI) {
-        context
-            .evm
-            .journaled_state
-            .initial_account_load(context.evm.env.block.coinbase, &[], &mut context.evm.db)
-            .map_err(EVMError::Database)?;
-    }
-
-    context.evm.load_access_list()?;
-    Ok(())
-}
-
-#[inline]
-pub fn deduct_caller<SPEC: Spec, EXT, DB: Database>(
-    context: &mut Context<EXT, DB>,
-) -> Result<(), EVMError<DB::Error>> {
-    // load caller's account.
-    let (caller_account, _) = context
-        .evm
-        .journaled_state
-        .load_account(context.evm.env.tx.caller, &mut context.evm.db)
-        .map_err(EVMError::Database)?;
-
-    // deduct gas cost from caller's account.
-    deduct_caller_inner::<SPEC>(caller_account, &context.evm.env);
-
-    Ok(())
-}
-
-#[inline]
-pub fn handle_reimburse_caller<SPEC: Spec, EXT, DB: Database>(
-    context: &mut Context<EXT, DB>,
-    gas: &Gas,
-) -> Result<(), EVMError<DB::Error>> {
-    let caller = context.evm.env.tx.caller;
-    let effective_gas_price = context.evm.env.effective_gas_price();
-
-    // return balance of not spend gas.
-    let (caller_account, _) = context
-        .evm
-        .journaled_state
-        .load_account(caller, &mut context.evm.db)
-        .map_err(EVMError::Database)?;
-
-    caller_account.info.balance = caller_account
-        .info
-        .balance
-        .saturating_add(effective_gas_price * U256::from(gas.remaining() + gas.refunded() as u64));
-
-    Ok(())
-}
-
-/// Reward beneficiary with gas fee.
-#[inline]
-pub fn reward_beneficiary<SPEC: Spec, EXT, DB: Database>(
-    context: &mut Context<EXT, DB>,
-    gas: &Gas,
-) -> Result<(), EVMError<DB::Error>> {
-    let beneficiary = context.evm.env.block.coinbase;
-    let effective_gas_price = context.evm.env.effective_gas_price();
-
-    // transfer fee to coinbase/beneficiary.
-    // EIP-1559 discard basefee for coinbase transfer. Basefee amount of gas is discarded.
-    let coinbase_gas_price = if SPEC::enabled(LONDON) {
-        effective_gas_price.saturating_sub(context.evm.env.block.basefee)
-    } else {
-        effective_gas_price
-    };
-
-    let (coinbase_account, _) = context
-        .evm
-        .journaled_state
-        .load_account(beneficiary, &mut context.evm.db)
-        .map_err(EVMError::Database)?;
-
-    coinbase_account.mark_touch();
-    coinbase_account.info.balance = coinbase_account
-        .info
-        .balance
-        .saturating_add(coinbase_gas_price * U256::from(gas.spend() - gas.refunded() as u64));
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use revm_interpreter::primitives::CancunSpec;
-
-    use super::*;
-
-    #[test]
-    fn test_consume_gas() {
-        let mut env = Env::default();
-        env.tx.gas_limit = 100;
-
-        let gas = handle_call_return::<CancunSpec>(&env, InstructionResult::Stop, Gas::new(90));
-        assert_eq!(gas.remaining(), 90);
-        assert_eq!(gas.spend(), 10);
-        assert_eq!(gas.refunded(), 0);
-    }
-
-    #[test]
-    fn test_consume_gas_with_refund() {
-        let mut env = Env::default();
-        env.tx.gas_limit = 100;
-
-        let mut return_gas = Gas::new(90);
-        return_gas.record_refund(30);
-
-        let gas = handle_call_return::<CancunSpec>(&env, InstructionResult::Stop, return_gas);
-        assert_eq!(gas.remaining(), 90);
-        assert_eq!(gas.spend(), 10);
-        assert_eq!(gas.refunded(), 30);
-
-        let gas = handle_call_return::<CancunSpec>(&env, InstructionResult::Revert, return_gas);
-        assert_eq!(gas.remaining(), 90);
-        assert_eq!(gas.spend(), 10);
-        assert_eq!(gas.refunded(), 0);
-    }
-
-    #[test]
-    fn test_revert_gas() {
-        let mut env = Env::default();
-        env.tx.gas_limit = 100;
-
-        let gas = handle_call_return::<CancunSpec>(&env, InstructionResult::Revert, Gas::new(90));
-        assert_eq!(gas.remaining(), 90);
-        assert_eq!(gas.spend(), 10);
-        assert_eq!(gas.refunded(), 0);
-    }
-}
+pub use main::{
+    deduct_caller_inner, main_deduct_caller, main_end, main_load, main_reimburse_caller,
+    main_return, main_reward_beneficiary,
+};
+pub use validation::{validate_env, validate_initial_tx_gas, validate_tx_against_state};
