+++ crates/revm/src/inspector/noop.rs
@@ -1,5 +1,5 @@
-use crate::{handler::register::GetInspector, Database, Inspector};
-
+use super::GetInspector;
+use crate::{Database, Inspector};
 /// Dummy [Inspector], helpful as standalone replacement.
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct NoOpInspector;
